requirements.txt
# =============================================================================
# Water Reminder Telegram Bot - Dependencies
# Python 3.13+ required
# =============================================================================

# Telegram Bot API
python-telegram-bot>=22.0

# Task Scheduling with persistent storage
APScheduler>=3.11.0

# Timezone support
pytz>=2025.0

# Environment variables management
python-dotenv>=1.0.0

# Database (SQLite via APScheduler jobstore)
sqlalchemy>=2.0.0

# =============================================================================
# Optional dependencies for development
# =============================================================================

# Testing (uncomment if needed)
# pytest>=8.0.0
# pytest-asyncio>=0.23.0

# Code quality (uncomment if needed)
# black>=24.0.0
# flake8>=7.0.0
# mypy>=1.8.0
